+++ packages/cli/CHANGELOG.md
@@ -14,6 +14,8 @@
 - We've improved our smart hashing for other use cases besides task running. The first improvement
   is that we now hash dependencies to determine whether to run a dependency install, or to skip!
   This is much more accurate than before, which only relied on lockfile modified timestamps.
+- Updated `moon docker` commands to take into account other programming languages when scaffolding
+  files.
 
 ##### Config